onnx: benchmark without a reference dump, and re-profile
A timing run measures time, and the time a convolution takes does not
depend on the numbers going through it. Requiring a full reference dump
for --benchmark meant regenerating hundreds of megabytes of promoted
intermediates before anyone could measure a kernel change, which is
enough friction to stop the measurement happening at all. The harness
now synthesises inputs from the shapes the graph declares, which needed
the model parser to keep them: OnnxValueInfo records the declared shape,
with a symbolic dimension — what a traced export writes for the batch —
kept as -1 rather than guessed.

The profile that unblocks contradicts the route the README and TODO
recorded as next. The slowest shapes are 1x1 convolutions with a large
channel count, 113 and 128 GFLOP/s, not the small-channel 3x3 layers,
which run at 195-277 while the deep 3x3 layers reach 349. The pattern is
reduction depth: a 1x1 reduces over 256 or 512 elements where a 3x3 over
the same channels reduces over nine times as many, so each output tile's
prologue and accumulator store are amortised nine times worse. Panel
size is not the difference — the fastest shape packs a 14 MB panel, far
past L2. A direct convolution avoiding the im2col expansion would have
been attacking shapes that are not the problem.

Also fixes two warnings the HEIC work introduced: a sign-extended
operand in the variable-width integer reader, and a null assigned to a
non-nullable EXIF map where an empty one was meant.

+++ dotnet/src/Xberg/Internal/Heif/HeifContainer.cs
@@ -427,7 +427,7 @@ internal static class HeifContainer
         long value = 0;
         for (int i = 0; i < size; i++)
         {
-            value = (value << 8) | (p < bytes.Length ? bytes[p] : 0);
+            value = (value << 8) | (p < bytes.Length ? bytes[p] : (byte)0);
             p++;
         }
         return value;
